suite système de logs

+++ demo/src/main/java/com/f1manager/demo/Formula1/Aileron/Ailerons.java
@@ -20,13 +20,13 @@ public class Ailerons {
     @Column(name="poid_aileron")
     private double poidsAileron;
     @Column(name="prix_aileron")
-    private double prixAileron;
+    private int prixAileron;
     @Column(name="niv_actuel")
     private int nivActuel;
     @Column(name = "image_aileron")
     private String imageAileron;
 
-    public Ailerons(double poidsAileron, double prixAileron, String imageAileron, int nivActuel) {
+    public Ailerons(double poidsAileron, int prixAileron, String imageAileron, int nivActuel) {
         this.poidsAileron = poidsAileron;
         this.prixAileron = prixAileron;
         this.imageAileron = imageAileron;
